# Remove an allocation from breaker reserve