The previous code captured `b`, which required a closure allocation on
each call of Reserve. Pre-allocating the closure as a field removes a
little bit of allocation from Reserve, which is worthwhile since Reserve
is used at high volume in performance-critical sections of the code.

Note: simpler options like defining `b.release` as a method rather than
a closure and returning  a method reference in `Reserve` still capture b, 
and therefore still allocate :-(.

/assign @markusthoemmes @vagababov 

<!--
Request Prow to automatically lint any go code in this PR:

/lint
-->


Benchmarks:
~~~~
benchmark                                 old ns/op     new ns/op     delta
BenchmarkBreakerReserve/sequential-16     119           92.6          -22.18%
BenchmarkBreakerReserve/parallel-16       202           161           -20.30%

benchmark                                 old allocs     new allocs     delta
BenchmarkBreakerReserve/sequential-16     1              0              -100.00%
BenchmarkBreakerReserve/parallel-16       1              0              -100.00%

benchmark                                 old bytes     new bytes     delta
BenchmarkBreakerReserve/sequential-16     16            0             -100.00%
BenchmarkBreakerReserve/parallel-16       16            0             -100.00%
~~~~